# backend_go/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkyViewLab Go Control Plane API",
    "version": "1.0.0-gate-b",
    "description": "Machine-checkable contract for the current Go /api/v1 control plane. This contract describes the implemented Gate B boundary; it does not claim that every workbench domain schema or production NO-GO item is complete. All JSON success responses use {data, requestId}; all JSON failures use {error:{code,message}, message, requestId}."
  },
  "servers": [
    {"url": "https://api.example.com", "description": "Production placeholder; deployment supplies the real origin"},
    {"url": "http://127.0.0.1:8080", "description": "Local development"}
  ],
...
  "paths": {
...
    "/api/v1/workbenches/{slug}/state": {
      "parameters": [{"$ref": "#/components/parameters/Slug"}],
      "get": {"operationId": "getWorkbenchState", "tags": ["workbenches"], "summary": "Get the current user's durable workbench draft", "description": "Returns revision 0 and an empty state when no draft exists. ETag and the body revision identify the same version.", "security": [{"sessionCookie": []}], "responses": {"200": {"$ref": "#/components/responses/WorkbenchState"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}},
      "put": {"operationId": "putWorkbenchState", "tags": ["workbenches"], "summary": "Compare-and-swap the current user's workbench draft", "description": "revision 0 creates an absent draft. A positive revision replaces exactly that revision; stale or concurrent writes return 409 and never overwrite newer state.", "security": [{"sessionCookie": []}], "parameters": [{"$ref": "#/components/parameters/Origin"}], "x-origin-required-for-cookie": true, "x-max-body-bytes": 4194304, "x-max-state-bytes": 1048576, "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchStateWriteRequest"}}}}, "responses": {"200": {"$ref": "#/components/responses/WorkbenchState"}, "400": {"$ref": "#/components/responses/Error400"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "409": {"$ref": "#/components/responses/Error409"}, "413": {"$ref": "#/components/responses/Error413"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
...
    "/api/v1/projects/{id}": {
      "parameters": [{"$ref": "#/components/parameters/ResourceID"}],
      "get": {"operationId": "getProject", "tags": ["projects"], "summary": "Get authorized project", "security": [{"sessionCookie": []}], "responses": {"200": {"$ref": "#/components/responses/Project"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}},
      "put": {"operationId": "replaceProject", "tags": ["projects"], "summary": "Replace project title and state", "security": [{"sessionCookie": []}], "parameters": [{"$ref": "#/components/parameters/Origin"}], "x-origin-required-for-cookie": true, "x-max-body-bytes": 4194304, "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectWriteRequest"}}}}, "responses": {"200": {"$ref": "#/components/responses/Project"}, "400": {"$ref": "#/components/responses/Error400"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}},
      "delete": {"operationId": "deleteProject", "tags": ["projects"], "summary": "Delete project", "security": [{"sessionCookie": []}], "parameters": [{"$ref": "#/components/parameters/Origin"}], "x-origin-required-for-cookie": true, "responses": {"200": {"$ref": "#/components/responses/Deleted"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "skyviewlab_session", "description": "Opaque HttpOnly BFF session. Browser unsafe operations also require an allowlisted Origin."},
      "oidcStateCookie": {"type": "apiKey", "in": "cookie", "name": "skyviewlab_oidc_state", "description": "Short-lived, one-time OIDC correlation cookie; not an authenticated application session."},
      "workerBearer": {"type": "http", "scheme": "bearer", "bearerFormat": "opaque", "description": "A distinct opaque token bound server-side to one worker ID and 1..64 exact tenantId/workspaceId/single-slug/action-allowlist scopes. Empty scopes, wildcards and duplicate boundaries are rejected. Production loads the bounded credential set from WORKER_CREDENTIALS_FILE; short-lived workload JWT or mTLS remains a deployment hardening boundary."}
    },
    "parameters": {
      "Origin": {"name": "Origin", "in": "header", "required": true, "description": "Exactly one browser Origin, matching the configured allowlist byte-for-byte.", "schema": {"type": "string", "format": "uri", "maxLength": 2048}},
...
      "Slug": {"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 128}},
      "ResourceID": {"name": "id", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 128}},
...
    },
    "schemas": {
      "ObjectMap": {"type": "object", "additionalProperties": true},
      "RequestID": {"type": "string", "minLength": 1, "maxLength": 128},
      "ErrorDetail": {"type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": {"code": {"type": "string", "minLength": 1}, "message": {"type": "string"}}},
      "ErrorEnvelope": {"type": "object", "additionalProperties": false, "required": ["error", "message", "requestId"], "properties": {"error": {"$ref": "#/components/schemas/ErrorDetail"}, "message": {"type": "string"}, "requestId": {"$ref": "#/components/schemas/RequestID"}}},
      "DataEnvelope": {"type": "object", "required": ["data", "requestId"], "properties": {"data": {}, "requestId": {"$ref": "#/components/schemas/RequestID"}}},
...
      "WorkbenchState": {"type": "object", "additionalProperties": false, "required": ["state", "revision"], "properties": {"state": {"$ref": "#/components/schemas/ObjectMap"}, "revision": {"type": "integer", "minimum": 0}, "updatedAt": {"type": "string", "format": "date-time"}}},
      "WorkbenchStateWriteRequest": {"type": "object", "additionalProperties": false, "required": ["state", "revision"], "properties": {"state": {"$ref": "#/components/schemas/ObjectMap"}, "revision": {"type": "integer", "minimum": 0}}},
      "Project": {"type": "object", "additionalProperties": false, "required": ["id", "tenantId", "workspaceId", "ownerUserId", "slug", "title", "state", "createdAt", "updatedAt"], "properties": {"id": {"type": "string"}, "tenantId": {"type": "string"}, "workspaceId": {"type": "string"}, "ownerUserId": {"type": "string"}, "slug": {"type": "string"}, "title": {"type": "string", "maxLength": 160}, "state": {"$ref": "#/components/schemas/ObjectMap"}, "createdAt": {"type": "string", "format": "date-time"}, "updatedAt": {"type": "string", "format": "date-time"}}},
...
      "ProjectWriteRequest": {"type": "object", "additionalProperties": false, "required": ["title", "state"], "properties": {"title": {"type": "string", "minLength": 1, "maxLength": 160}, "state": {"$ref": "#/components/schemas/ObjectMap"}}},
...
      "Deleted": {"type": "object", "additionalProperties": false, "required": ["deleted"], "properties": {"deleted": {"const": true}}},
...
      "WorkbenchStateEnvelope": {"allOf": [{"$ref": "#/components/schemas/DataEnvelope"}, {"type": "object", "properties": {"data": {"$ref": "#/components/schemas/WorkbenchState"}}}]},
      "ProjectEnvelope": {"allOf": [{"$ref": "#/components/schemas/DataEnvelope"}, {"type": "object", "properties": {"data": {"$ref": "#/components/schemas/Project"}}}]},
...
      "DeletedEnvelope": {"allOf": [{"$ref": "#/components/schemas/DataEnvelope"}, {"type": "object", "properties": {"data": {"$ref": "#/components/schemas/Deleted"}}}]}
    },
    "responses": {
...
      "WorkbenchState": {"description": "Durable per-user workbench draft and CAS revision", "headers": {"ETag": {"description": "Strong revision tag matching data.revision", "schema": {"type": "string", "pattern": "^\\\"workbench-state-[0-9]+\\\"$"}}, "Cache-Control": {"schema": {"type": "string", "const": "private, no-store"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkbenchStateEnvelope"}}}},
      "Project": {"description": "Project", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectEnvelope"}}}},
...
      "Deleted": {"description": "Resource deleted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeletedEnvelope"}}}},
...
      "Error400": {"description": "Invalid request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error401": {"description": "Authentication or correlation failed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error403": {"description": "Origin, role, or resource authorization denied", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error404": {"description": "Resource not found within the authorized scope", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error409": {"description": "State, revision, idempotency, or immutable-workflow conflict", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error413": {"description": "Request or encoded job input exceeds the documented limit", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
      "Error500": {"description": "Internal persistence, encoding, or audit failure", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
      "Error503": {"description": "Required identity, authorization, Worker, database, or compute dependency is unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
    }
  }
}